requirements.txt
anthropic==0.85.0
feedparser==6.0.12
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
python-dotenv==1.2.2
requests==2.32.3
requests-oauthlib==2.0.0
weasyprint==68.1
